--- src/budgets/cli/__init__.py
@@ -1,1 +1,0 @@
-""" Cmd interface for Budgets """
