--- src/budgets/core/tests/__init__.py
@@ -1,2 +1,0 @@
-from budgets.core.tests.test_budgets import TestBudgets
-from budgets.core.tests.test_wavefront_error import TestWaveFrontError
