+++ src/budgie/cli/__init__.py
@@ -1,0 +1,1 @@
+""" Cmd interface for Budgie """
