+++ src/budgie/core/tests/__init__.py
@@ -1,0 +1,2 @@
+from budgie.core.tests.test_budgets import TestBudgets
+from budgie.core.tests.test_wavefront_error import TestWaveFrontError
